+++ src/os2datascanner/projects/report/reportapp/tests/test_utils.py
@@ -2,27 +2,7 @@ import pytest
 from django.db import IntegrityError
 
 from ...organizations.models import Organization, Account, Alias, AliasType
-from ..utils import get_or_create_user_aliases
 from ..views.utilities.document_report_utilities import is_owner
-
-
-@pytest.fixture(scope="class")
-def saml_user_data():
-    user_data = {
-        'email': 'sam_single@saml.com',
-        'username': 'Sam',
-        'first_name': 'Sam Single',
-        'last_name': 'Sign-On',
-        'sid': 'S-DIG',
-        'user_identity': {
-            'email': ['sam_single@saml.com'],
-            'username': ['Sam'],
-            'first_name': ['Sam Single'],
-            'surname': ['Sign-On'],
-            'sid': ['S-DIG']
-        }
-    }
-    return user_data
 
 
 @pytest.mark.django_db
@@ -47,115 +27,6 @@ class TestUtils:
                                               organization=org)
         cls.account_jack = account_jack
         cls.user_jack = account_jack.user
-
-    def test_email_alias_is_created(self, saml_user_data):
-        # Arrange / Act
-        get_or_create_user_aliases(saml_user_data)
-
-        # Assert
-        assert Alias.objects.filter(
-            _value="sam_single@saml.com", _alias_type=AliasType.EMAIL).exists() is True
-
-    def test_sid_alias_is_created(self, saml_user_data):
-        # Arrange / Act
-        get_or_create_user_aliases(saml_user_data)
-
-        # Assert
-        assert Alias.objects.filter(
-            _value="S-DIG", _alias_type=AliasType.SID).exists() is True
-
-    def test_aliases_are_related_to_user(self, saml_user_data):
-        # Arrange / Act
-        get_or_create_user_aliases(saml_user_data)
-
-        # Assert
-        assert self.user_sam.aliases.count() == 2
-
-    def test_aliases_are_related_to_account(self, saml_user_data):
-        # Arrange / Act
-        get_or_create_user_aliases(saml_user_data)
-
-        # Assert
-        assert self.account_sam.aliases.count() == 2
-
-    def test_existing_alias_doesnt_duplicate(self, saml_user_data):
-        # Arrange
-        Alias.objects.create(
-            user=self.user_sam,
-            account=self.account_sam,
-            _alias_type=AliasType.SID,
-            _value="S-DIG"
-        )
-        Alias.objects.create(
-            user=self.user_sam,
-            account=self.account_sam,
-            _alias_type=AliasType.EMAIL,
-            _value="sam_single@saml.com"
-        )
-
-        # Act
-        get_or_create_user_aliases(saml_user_data)
-
-        # Assert
-        assert self.user_sam.aliases.count() == 2
-
-    def test_existing_aliases_only_related_to_user(self, saml_user_data):
-        # Arrange
-        Alias(
-            user=self.user_sam,
-            _alias_type=AliasType.SID,
-            _value="S-DIG"
-        ).save(prevent_mismatch=False)  # Needed to force creation of alias without account
-        Alias(
-            user=self.user_sam,
-            _alias_type=AliasType.EMAIL,
-            _value="sam_single@saml.com"
-        ).save(prevent_mismatch=False)  # Needed to force creation of alias without account
-
-        # Act
-        get_or_create_user_aliases(saml_user_data)
-
-        # Assert
-        # These aliases should now be connected to the Account object as well.
-        # I.e. there should only be two aliases; not four
-        assert self.user_sam.aliases.count() == 2
-
-    def test_existing_aliases_only_related_to_account(self, saml_user_data):
-        # This isn't possible due to DB constraints - Should raise IntegrityError
-        with pytest.raises(IntegrityError):  # "Assert"
-            # Arrange
-            Alias.objects.create(
-                account=self.account_sam,
-                _alias_type=AliasType.SID,
-                _value="S-DIG"
-            )
-            Alias.objects.create(
-                acccount=self.account_sam,
-                _alias_type=AliasType.EMAIL,
-                _value="sam_single@saml.com"
-            )
-            # Act
-            get_or_create_user_aliases(saml_user_data)
-
-    def test_user_and_user_acccount_alias_exists_will_clean_up(self, saml_user_data):
-        # Arrange
-        #  We could potentially see a situation, where both the aliases
-        #  Alias(user=A, account=A.account) and Alias(user=A) exist.
-        Alias.objects.create(
-            user=self.user_sam,
-            account=self.account_sam,
-            _alias_type=AliasType.SID,
-            _value="S-DIG"
-        )
-        Alias(
-            user=self.user_sam,
-            _alias_type=AliasType.SID,
-            _value="S-DIG"
-        ).save(prevent_mismatch=False)  # Needed to force creation of alias without account
-        # Act
-        get_or_create_user_aliases(saml_user_data)
-        # Assert
-        assert Alias.objects.filter(_alias_type=AliasType.SID).count() == 1
 
     def test_creating_alias_with_no_account(self):
         """Creating an alias with no account should throw an exception."""
@@ -281,22 +152,7 @@ class TestUtils:
             # Act
             mismatched_alias.save()
 
-    def test_alias_and_report_owner_case_mismatch(self, saml_user_data):
-        """
-        Test that a potential case mismatch between Account.alias and DocumentReport.owner won't
-        prevent them from getting correctly paired up.
-        """
-
-        # Arrange
-        file_owner = saml_user_data["email"]
-        reversed_case_value = file_owner.swapcase()
-
-        get_or_create_user_aliases(saml_user_data)
-
-        # Act/assert
-        assert is_owner(reversed_case_value, self.account_sam)
-
-    def test_alias_and_report_owner_no_match(self, saml_user_data):
+    def test_alias_and_report_owner_no_match(self):
         """
         Test that the is_owner returns False when DocumentReport.owner doesn't match any alias for
         the given account.